requirements.txt
fastapi
uvicorn
pydub
python-multipart
SpeechRecognition
openai
python-dotenv
Boto3
motor
pymongo[srv]
librosa